requirements.txt
pandas
yfinance
matplotlib
jinja2
icalendar
pyyaml
beautifulsoup4
requests
lxml
recurring-ical-events
openai
python-dotenv
pyyaml
